Fix E2E test: Create plugin data directory if missing
The E2E test was failing because the plugin data directory didn't exist
when using --skip-build. The workflow should create this directory, but
if it doesn't exist, the E2E test should create it to be more robust.

This fixes the error: Plugin data directory not found, plugin may not be installed